+++ Plugins/Events/Events/ItemEvents.hpp
@@ -20,6 +20,7 @@ private:
     static void CloseInventoryHook(CNWSItem*, NWNXLib::API::Types::ObjectID, int32_t);
     static uint32_t FindItemWithBaseItemIdHook(CItemRepository*, uint32_t, int32_t);
     static int32_t LearnScrollHook(CNWSCreature*, NWNXLib::API::Types::ObjectID);
+    static int32_t CanEquipItemHook(CNWSCreature*, CNWSItem*, uint32_t*, int32_t, int32_t, int32_t, CNWSPlayer*);
     static int32_t RunEquipHook(CNWSCreature*, NWNXLib::API::Types::ObjectID, uint32_t, uint32_t);
     static int32_t RunUnequipHook(CNWSCreature*, NWNXLib::API::Types::ObjectID, NWNXLib::API::Types::ObjectID, uint8_t, uint8_t, int32_t, uint32_t);
     static void ItemEventHandlerHook(CNWSItem*, uint32_t, NWNXLib::API::Types::ObjectID, void*, uint32_t, uint32_t);