+++ CHANGELOG.md
@@ -9,7 +9,8 @@ https://github.com/nwnxee/unified/compare/build8193.10...HEAD
 
 ### Added
 - Events: Added Keyboard Input event to InputEvents
-- Events: Added Use Item validation event to allow control of the red highlighting behaviour for items that cannot be used.
+- Events: Added Use Item validation event to allow control of item use restrictions, and unusable red inventory highlight.
+- Events: Added Equip Item validation event to allow control of item equip restrictions.
 
 ##### New Plugins
 N/A